Split McpConnectionManager in two.
stdio MCP servers should have their lifecycle tied to the
harness' rather than an individual thread's. It's hard to
make that change when the global connection/spawning
responsibilities are mixed with the per-thread responsibilities.

This change leaves the McpConnectionManager API intact so
that client changes can be reviewed separately in a follow-up.
